Add files via upload
This is just the game list which calls a request in a request, and it posts my own json object with only what we want. I'm getting my computer set up again right now but I am gonna work on this during work tomorrow and when I get home.  But yeah this is the data that will be returned to our html.  This would be for the profile subtab "myGames"